Update to `bevy_replicon` 0.32
Also improve docs.

+++ src/client.rs
@@ -88,7 +88,7 @@ impl RepliconRenetClientPlugin {
                 "forwarding {} sent bytes over channel {channel_id}",
                 message.len()
             );
-            renet_client.send_message(channel_id, message)
+            renet_client.send_message(channel_id as u8, message)
         }
     }
 }
